Fix the GBM_MICRO macro
The version of gbm can sometimes be suffixed with ‘-devel’. This was
making the GBM_MICRO define come out as 0-devel which was generating a
warning when it was used in a #if check. This patch makes it chop off
anything after a ‘-’ using sed.

Reviewed-by: Robert Bragg <robert@linux.intel.com>

(cherry picked from commit 64472027fb5750971e94c0b4c6c624a39e5abe2f)