loginDialog: handle typing at screenshield
If the screenshield is up and the user starts typing, we need to
keep track of it and replay it later, otherwise what they type
will be ignored.

This is useful when the user has disable-user-list set to TRUE,
and will also be useful to bring feature parity between the
login screen and unlock screen.

https://bugzilla.gnome.org/show_bug.cgi?id=702308