cogl-path: Generate the primitive before using it for clipping
The data->fill_primitive member is meant to be generated on demand in
an accessor function. However when using it for clipping the code was
trying to directly use it without calling the accessor so it would
crash if nothing else caused the primitive to be generated. This was
making the texture-fbo test in Clutter crash.

(cherry picked from commit 0e913664339739812848cdb73bec64c066860e65)